Removed nego struct from direct access.
